[app] Cleaning up types and app structure
