--- app/src/chatBro/queries/TopFriends/types.d.ts
@@ -1,17 +1,0 @@
-declare namespace TopFriendsTypes {
-  interface Filters {
-    limit?: number;
-    contact?: string;
-    word?: string;
-    groupChat?: string;
-  }
-
-  interface ChartData {
-    friend: string;
-    total: number;
-    sent: number;
-    received: number;
-  }
-
-  type Results = ChartData[];
-}
