+++ app/src/chatBro/queries/TopFriends/types.ts
@@ -1,0 +1,15 @@
+export interface ITopFriendsFilters {
+  limit?: number;
+  contact?: string;
+  word?: string;
+  groupChat?: string;
+}
+
+export interface ITopFriendsChartData {
+  friend: string;
+  total: number;
+  sent: number;
+  received: number;
+}
+
+export type TTopFriendsResults = ITopFriendsChartData[];
