--- app/src/components/css.d.ts
@@ -1,9 +1,0 @@
-declare module '*.scss' {
-  const content: { [className: string]: string };
-  export default content;
-}
-
-declare module '*.css' {
-  const content: { [className: string]: string };
-  export default content;
-}
