--- app/src/utils/initUtils/contacts/types.d.ts
@@ -1,7 +1,0 @@
-declare namespace ContactOptions {
-  interface ContactData {
-    value?: string;
-    label?: string;
-    mycount?: number;
-  }
-}
