+++ app/src/utils/initUtils/contacts/types.ts
@@ -1,0 +1,5 @@
+export interface IContactData {
+  value?: string;
+  label?: string;
+  mycount?: number;
+}
